Move GCP Video Intelligence to core
