+++ docs/howto/operator/gcp/video.rst
@@ -30,21 +30,21 @@ CloudVideoIntelligenceDetectVideoLabelsOperator
 Performs video annotation, annotating video labels.
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_video_intelligence_operator.CloudVideoIntelligenceDetectVideoLabelsOperator`
+:class:`airflow.gcp.operators.video_intelligence.CloudVideoIntelligenceDetectVideoLabelsOperator`
 
 Arguments
 """""""""
 
 Some arguments in the example DAG are taken from the OS environment variables:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_os_args]
       :end-before: [END howto_operator_video_intelligence_os_args]
 
 Input uri is an uri to a file in Google Cloud Storage
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_other_args]
       :end-before: [END howto_operator_video_intelligence_other_args]
@@ -52,7 +52,7 @@ Input uri is an uri to a file in Google Cloud Storage
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_labels]
@@ -60,7 +60,7 @@ Using the operator
 
 You can use the annotation output via Xcom:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_labels_result]
@@ -69,7 +69,7 @@ You can use the annotation output via Xcom:
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_video_intelligence_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/video_intelligence.py
     :language: python
     :dedent: 4
     :start-after: [START gcp_video_intelligence_detect_labels_template_fields]
@@ -89,21 +89,21 @@ CloudVideoIntelligenceDetectVideoExplicitContentOperator
 Performs video annotation, annotating explicit content.
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_video_intelligence_operator.CloudVideoIntelligenceDetectVideoExplicitContentOperator`
+:class:`airflow.gcp.operators.video_intelligence.CloudVideoIntelligenceDetectVideoExplicitContentOperator`
 
 Arguments
 """""""""
 
 Some arguments in the example DAG are taken from the OS environment variables:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_os_args]
       :end-before: [END howto_operator_video_intelligence_os_args]
 
 Input uri is an uri to a file in Google Cloud Storage
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_other_args]
       :end-before: [END howto_operator_video_intelligence_other_args]
@@ -111,7 +111,7 @@ Input uri is an uri to a file in Google Cloud Storage
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_explicit_content]
@@ -119,7 +119,7 @@ Using the operator
 
 You can use the annotation output via Xcom:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_explicit_content_result]
@@ -128,7 +128,7 @@ You can use the annotation output via Xcom:
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_video_intelligence_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/video_intelligence.py
     :language: python
     :dedent: 4
     :start-after: [START gcp_video_intelligence_detect_explicit_content_template_fields]
@@ -148,21 +148,21 @@ CloudVideoIntelligenceDetectVideoShotsOperator
 Performs video annotation, annotating explicit content.
 
 For parameter definition, take a look at
-:class:`airflow.contrib.operators.gcp_video_intelligence_operator.CloudVideoIntelligenceDetectVideoShotsOperator`
+:class:`airflow.gcp.operators.video_intelligence.CloudVideoIntelligenceDetectVideoShotsOperator`
 
 Arguments
 """""""""
 
 Some arguments in the example DAG are taken from the OS environment variables:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_os_args]
       :end-before: [END howto_operator_video_intelligence_os_args]
 
 Input uri is an uri to a file in Google Cloud Storage
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :start-after: [START howto_operator_video_intelligence_other_args]
       :end-before: [END howto_operator_video_intelligence_other_args]
@@ -170,7 +170,7 @@ Input uri is an uri to a file in Google Cloud Storage
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_video_shots]
@@ -178,7 +178,7 @@ Using the operator
 
 You can use the annotation output via Xcom:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_video_intelligence.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_video_intelligence.py
       :language: python
       :dedent: 4
       :start-after: [START howto_operator_video_intelligence_detect_video_shots_result]
@@ -187,7 +187,7 @@ You can use the annotation output via Xcom:
 Templating
 """"""""""
 
-.. literalinclude:: ../../../../airflow/contrib/operators/gcp_video_intelligence_operator.py
+.. literalinclude:: ../../../../airflow/gcp/operators/video_intelligence.py
     :language: python
     :dedent: 4
     :start-after: [START gcp_video_intelligence_detect_video_shots_template_fields]
