+++ tests/gcp/operators/test_video_intelligence_system_helper.py
